+++ saleor/core/utils/__init__.py
@@ -1,5 +1,6 @@
 import decimal
 import logging
+import socket
 from json import JSONEncoder
 from urllib.parse import urljoin
 
@@ -51,10 +52,38 @@ def build_absolute_uri(location):
     return iri_to_uri(location)
 
 
+def is_valid_ipv4(ip: str) -> bool:
+    """Check whether the passed IP is a valid V4 IP address."""
+    try:
+        socket.inet_pton(socket.AF_INET, ip)
+    except socket.error:
+        return False
+    return True
+
+
+def is_valid_ipv6(ip: str) -> bool:
+    """Check whether the passed IP is a valid V6 IP address."""
+    try:
+        socket.inet_pton(socket.AF_INET6, ip)
+    except socket.error:
+        return False
+    return True
+
+
 def get_client_ip(request):
-    ip = request.META.get("HTTP_X_FORWARDED_FOR", None)
-    if ip:
-        return ip.split(",")[0].strip()
+    """Retrieve the IP address from the request data.
+
+    Tries to get a valid IP address from X-Forwarded-For, if the user is hiding behind
+    a transparent proxy or if the server is behind a proxy.
+
+    If no forwarded IP was provided or all of them are invalid,
+    it fallback to the requester IP.
+    """
+    ip = request.META.get("HTTP_X_FORWARDED_FOR", "")
+    ips = ip.split(",")
+    for ip in ips:
+        if is_valid_ipv4(ip) or is_valid_ipv6(ip):
+            return ip
     return request.META.get("REMOTE_ADDR", None)